# improve hashing performance w/ specialized hashers 