--- a/.github/workflows/benchmark.yaml
+++ b/.github/workflows/benchmark.yaml
@@ -35,6 +35,8 @@ jobs:
           cd tests/db-benchmark
           Rscript -e 'install.packages("data.table", repos="https://Rdatatable.github.io/data.table")'
           Rscript groupby-datagen.R 1e7 1e2 5 0
+          echo "ON STRINGS"
           python main.py on_strings
+          echo "ON CATEGORICALS"
           python main.py